requirements.txt
google-ads
pandas
progressbar2